hxdmp: -help and -length flags.
You can now specify the display length of the data using '-l', '-length' flags. '-h' should now display the help message. Formatting of the output was changed to be more customizable with variables. Makefile now has a 'debug' rule for building program in debug mode, running a debugger and removing the debugger output.